+++ frontend/src/components/StudentsTable/types.ts
