+++ frontend/src/pages/home.tsx
@@ -1,4 +1,4 @@
-import StudentsTable from "../components/StudentTables/StudentsTable";
+import StudentsTable from "../components/StudentsTable/StudentsTable";
 
 export default function Home() {
   return <StudentsTable />;
