--- frontend/src/sampleStudentData 2.json
@@ -1,29 +1,0 @@
-{
-  "student": {
-    "firstName": "John",
-    "lastName": "Doe",
-    "email": "John@aol.com",
-    "phoneNumber": "1005882300"
-  },
-  "emergency": {
-    "firstName": "Bob",
-    "lastName": "James",
-    "email": "bob@aol.com",
-    "phoneNumber": "2005882300"
-  },
-  "serviceCoordinator": {
-    "firstName": "William",
-    "lastName": "John",
-    "email": "William@aol.com",
-    "phoneNumber": "3005882300"
-  },
-  "location": "HERE",
-  "medication": "ADVIL",
-  "birthday": "2023-10-01T00:00Z",
-  "intakeDate": "2023-10-03T00:00Z",
-  "tourDate": "2023-10-05T00:00Z",
-  "prog1": ["intro"],
-  "prog2": ["tds", "sdp"],
-  "dietary": ["seafood", "dairy"],
-  "otherString": "peanuts"
-}
